Refactor reservoir layers structure
This commit reorganizes the reservoir layers by moving the BaseReservoir and its related classes into a new 'layers' subdirectory inside 'reservoirs'.

Similarily the same with the BaseCell class and its related classes to the 'cells' directory inside 'reservoirs'.

Necessary refactor with other related modules was made also.

The changes enhance the modularity and clarity of the reservoir components, facilitating easier maintenance and extension in future developments.