Add dendrocycle kernel and input initializers.
A dendrocycle is a graph that has a cycle and legs of nodes coming out of it
like dendrites.

This topology makes the cycle encode the current states of the system
and the legs to function as discrete delay embeddings of the past.

The dendrocycle input maps the input vector only to the cycle neurons,
making the input only drive the current encoded state of the system.